+++ src/App.svelte
@@ -9,16 +9,28 @@
   SCR_CONFIG_STORE.setConsoleLogStores(false);
   SCR_CONFIG_STORE.setNavigationHistoryLimit(10);
   SCR_CONFIG_STORE.setHashMode(true);
+  SCR_CONFIG_STORE.setUseScroll(true);
+  SCR_CONFIG_STORE.setScrollProps({
+    top: 0,
+    left: 0,
+    behavior: "smooth",
+    timeout: 10,
+  });
   SCR_CONFIG_STORE.setOnError((err, routeObjParams) => {
     console.log("GLOBAL ERROR CONFIG", routeObjParams);
   });
   SCR_CONFIG_STORE.setBeforeEnter([
-    (resolve) => {
+    (resolve, routeFrom, routeTo, routeObjParams, payload) => {
+      payload.firstGBEF = "set on Global Enter Function - 1!"
+      console.log(payload)
       console.log("GBER-1");
       resolve(true);
     },
-    (resolve) => {
+    (resolve, routeFrom, routeTo, routeObjParams, payload) => {
+      console.log(payload)
+      payload.secondGBEF = "set on Global Enter Function! - 2"
       console.log("GBER-2");
+      console.log(payload)
       resolve(true);
     },
   ]);
@@ -28,13 +40,15 @@
       name: "rootRoute",
       path: "/svelte-client-router",
       lazyLoadComponent: () => import("./testComponents/SCR_Root.svelte"),
-      lazyLoadLayoutComponent: () => import("./testComponents/SCR_Layout.svelte"),
+      lazyLoadLayoutComponent: () =>
+        import("./testComponents/SCR_Layout.svelte"),
       title: "SCR - Root Route",
     },
     {
       name: "routeOne",
       path: "/svelte-client-router/test1",
       component: SCR_C1,
+      ignoreScroll: true,
       beforeEnter: [
         (resolve) => {
           console.log("beforeEnter Executed");
@@ -42,7 +56,10 @@
         },
         (resolve) => {
           console.log("beforeEnter Executed2");
-          setTimeout(() => resolve({ redirect: "/svelte-client-router" }), 1000);
+          setTimeout(
+            () => resolve({ redirect: "/svelte-client-router" }),
+            1000
+          );
         },
       ],
       title: "First Route Title",
@@ -54,13 +71,22 @@
       lazyLoadComponent: () => import("./testComponents/SCR_C2.svelte"),
       title: "Second Route Title",
       beforeEnter: [
-        (resolve, rFrom, rTo, params) => {
+        (resolve, rFrom, rTo, params, payload) => {
+          console.log(payload);
+          payload.myCustomParam = "123";
+          console.log("<<<<>>>>>");
           console.log("beforeEnter Executed");
           console.log(params);
           setTimeout(() => resolve(true), 1000);
         },
       ],
       loadingProps: { loadingText: "Carregando 2..." },
+      scrollProps: {
+        top: 100,
+        left: 100,
+        behavior: "smooth",
+        timeout: 1000,
+      },
     },
     {
       name: "routeThree",
@@ -85,10 +111,18 @@
         myCustomParam: "This Param was set in the Router Definition",
       },
       title: "Four Route Title",
-      lazyLoadComponent: () => import('./testComponents/SCR_C4.svelte'),
-      lazyLoadLayoutComponent: () => import("./testComponents/SCR_Layout_Global.svelte"),
+      lazyLoadComponent: () => import("./testComponents/SCR_C4.svelte"),
+      lazyLoadLayoutComponent: () =>
+        import("./testComponents/SCR_Layout_Global.svelte"),
+      beforeEnter: [
+        (resolve, rFrom, rTo, params, payload) => {
+          console.log(payload);
+          payload.myCustomParam2 = "set on Route Before Enter"
+          resolve(true)
+        },
+      ],
       afterBeforeEnter: (routeObjParams) => {
-        console.log("After BE Route Four")
+        console.log("After BE Route Four");
         console.log(routeObjParams);
       },
       loadingProps: { loadingText: "Carregando 4..." },
@@ -97,13 +131,16 @@
       name: "routeFive",
       path: "/svelte-client-router/test5",
       title: "Five Route Title",
-      lazyLoadComponent: () => import('./testComponents/SCR_C5.svelte'),
+      lazyLoadComponent: () => import("./testComponents/SCR_C5.svelte"),
       loadingProps: { loadingText: "Carregando 5..." },
       ignoreLayout: true,
-      ignoreGlobalBeforeFunction: true
+      ignoreGlobalBeforeFunction: true,
     },
   ];
-
 </script>
 
-<SCR_ROUTER_COMPONENT bind:routes defaultLayoutComponent={SCR_Layout_Global} allProps={{ allPropsTest: "Passing To ALL"}} />
+<SCR_ROUTER_COMPONENT
+  bind:routes
+  defaultLayoutComponent={SCR_Layout_Global}
+  allProps={{ allPropsTest: "Passing To ALL" }}
+/>